requirements.txt
openpyxl
pk4adi >=0.1.3
pandas
PyQt5-Frameless-Window>=0.2.7
PyQt_Fluent_Widgets